[feat] minor changes, Comment, Issue, Project, User.java into entity package
