+++ src/main/java/com/example/IssueTrackingSystem/domain/entity/Comment.java
@@ -1,7 +1,5 @@
-package com.example.IssueTrackingSystem.domain;
+package com.example.IssueTrackingSystem.domain.entity;
 
-import com.example.IssueTrackingSystem.domain.common.BaseEntity;
-import com.example.IssueTrackingSystem.domain.enums.UserRole;
 import jakarta.persistence.*;
 import lombok.*;
 import org.hibernate.annotations.DynamicInsert;
@@ -14,21 +12,21 @@ import org.hibernate.annotations.DynamicUpdate;
 @DynamicInsert
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 @AllArgsConstructor
-public class Project extends BaseEntity {
-
+public class Comment {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long projectId;
-
-    @Column(nullable = false, length = 20)
-    private String title;
+    private Long commentId;
 
     @Column(columnDefinition = "TEXT")
-    private String content;
+    private String comment;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "user_id")
     private User user;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "issue_id")
+    private Issue issue;
 
     public void setUser(User user) {
 //        // 기존에 이미 등록되어 있던 관계를 제거
@@ -44,4 +42,17 @@ public class Project extends BaseEntity {
 //        }
     }
 
+    public void setIssue(Issue issue) {
+//        // 기존에 이미 등록되어 있던 관계를 제거
+//        if (this.user != null) {
+//            this.user.getQuestionList().remove(this);
+//        }
+
+        this.issue = issue;
+
+//        // 양방향 관계를 설정
+//        if (user != null) {
+//            user.getQuestionList().add(this);
+//        }
+    }
 }
