+++ src/main/java/com/example/IssueTrackingSystem/domain/entity/Issue.java
@@ -1,5 +1,6 @@
-package com.example.IssueTrackingSystem.domain;
+package com.example.IssueTrackingSystem.domain.entity;
 
+import com.example.IssueTrackingSystem.domain.common.BaseEntity;
 import com.example.IssueTrackingSystem.domain.enums.IssueStatus;
 import jakarta.persistence.*;
 import lombok.*;
@@ -13,21 +14,25 @@ import org.hibernate.annotations.DynamicUpdate;
 @DynamicInsert
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 @AllArgsConstructor
-public class Comment {
+public class Issue extends BaseEntity {
+
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long commentId;
+    private Long issueId;
+
+    @Column(nullable = false, length = 20)
+    private String title;
 
     @Column(columnDefinition = "TEXT")
-    private String comment;
+    private String description;
+
+    @Enumerated(EnumType.STRING)
+    @Column(columnDefinition = "VARCHAR(15) DEFAULT 'DEFAULT'")
+    private IssueStatus issueStatus;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "user_id")
     private User user;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "issue_id")
-    private Issue issue;
 
     public void setUser(User user) {
 //        // 기존에 이미 등록되어 있던 관계를 제거
@@ -36,20 +41,6 @@ public class Comment {
 //        }
 
         this.user = user;
-
-//        // 양방향 관계를 설정
-//        if (user != null) {
-//            user.getQuestionList().add(this);
-//        }
-    }
-
-    public void setIssue(Issue issue) {
-//        // 기존에 이미 등록되어 있던 관계를 제거
-//        if (this.user != null) {
-//            this.user.getQuestionList().remove(this);
-//        }
-
-        this.issue = issue;
 
 //        // 양방향 관계를 설정
 //        if (user != null) {
