+++ src/main/java/com/example/IssueTrackingSystem/repository/UserRepository.java
@@ -1,12 +1,12 @@
 package com.example.IssueTrackingSystem.repository;
 
-import com.example.IssueTrackingSystem.domain.User;
+import com.example.IssueTrackingSystem.domain.entity.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
-
-import java.util.List;
 
 @Repository
 public interface UserRepository extends JpaRepository<User, Long> {
     User findByUserName(String userName);
 }
+
+
