+++ src/main/java/com/example/IssueTrackingSystem/service/UserService/UserQueryService.java
@@ -1,6 +1,6 @@
 package com.example.IssueTrackingSystem.service.UserService;
 
-import com.example.IssueTrackingSystem.domain.User;
+import com.example.IssueTrackingSystem.domain.entity.User;
 import com.example.IssueTrackingSystem.web.dto.UserRequestDTO;
 
 public interface UserQueryService {
