+++ src/main/java/com/example/IssueTrackingSystem/service/UserService/UserQueryServiceImpl.java
@@ -1,6 +1,6 @@
 package com.example.IssueTrackingSystem.service.UserService;
 
-import com.example.IssueTrackingSystem.domain.User;
+import com.example.IssueTrackingSystem.domain.entity.User;
 import com.example.IssueTrackingSystem.repository.UserRepository;
 import com.example.IssueTrackingSystem.web.dto.UserRequestDTO;
 import lombok.RequiredArgsConstructor;
